Extract `TraceDeserializer` from `RdProcessRunner`
